Changed

+++ src/BmiService.java
@@ -1,5 +1,5 @@
 public class BmiService {
-    public double calculate(int kg, double m) {
+    public int calculate(int kg, double m) {
         double bmiIndex = (kg / (m * m));
         return (int) bmiIndex;
     }
